FastRicianLikelihoods

Documentation for FastRicianLikelihoods.

FastRicianLikelihoods.Rice

FastRicianLikelihoods.besseli0m1x

FastRicianLikelihoods.besseli1i0

FastRicianLikelihoods.besseli2

FastRicianLikelihoods.besseli2x

FastRicianLikelihoods.Rice — Type

Rice(ν, σ)

The Rice distribution with parameters ν and σ has probability density function:

\[f(x; \nu, \sigma) = \frac{x}{\sigma^2} \exp\left( \frac{-(x^2 + \nu^2)}{2\sigma^2} \right) I_0\left( \frac{x\nu}{\sigma^2} \right).\]

External links:

Rice distribution on Wikipedia

FastRicianLikelihoods.besseli0m1x — Method

besseli0m1x(x::T) where {T <: Union{Float32, Float64}}

Scaled modified Bessel function of the first kind of order zero minus one, $(I_0(x) - 1)*e^{-x}$.

FastRicianLikelihoods.besseli1i0 — Method

besseli1i0(x::T) where {T <: Union{Float32, Float64}}

Ratio of modified Bessel functions of the first kind of orders one and zero, $I_1(x) / I_0(x)$.

FastRicianLikelihoods.besseli2 — Method

besseli2(x::T) where {T <: Union{Float32, Float64}}

Modified Bessel function of the first kind of order two, $I_2(x)$.

FastRicianLikelihoods.besseli2x — Method

besseli2x(x::T) where {T <: Union{Float32, Float64}}

Scaled modified Bessel function of the first kind of order two, $I_2(x)*e^{-|x|}$.